added stuff

+++ client/nextjs-codec/app/api/rooms/route.ts
@@ -1,34 +1,45 @@
-import { NextResponse } from 'next/server';
+import { NextRequest, NextResponse } from 'next/server';
 import dbConnect from '../../../lib/dbConnect';
 import mongoose from 'mongoose';
 import { ObjectId } from 'mongodb';
 
-export async function DELETE(request: Request) {
-  const { searchParams } = new URL(request.url);
+export async function DELETE(request: NextRequest) {
+  console.log('DELETE request received');
 
-  const room_id: string = searchParams.get('room_id')!;
+  const searchParams = request.nextUrl.searchParams;
+  const room_id: string | null = searchParams.get('room_id');
+
+  console.log('room_id:', room_id);
+
+  if (!room_id) {
+    return new NextResponse('Room ID is required', { status: 400 });
+  }
 
   try {
     await dbConnect();
-
+    console.log('Database connected successfully');
+    
     const db = mongoose.connection;
-
     const rooms = db.collection('coderooms');
 
-    rooms.deleteOne({ _id: new ObjectId(room_id) });
-    return new NextResponse(
-      `Room with id of: ${room_id} was successfully deleted`,
-      {
-        status: 200,
-      }
-    );
-  } catch (e) {
-    return NextResponse.json(
-      {
-        message: `Failed to do deletion operation on room with id of: ${room_id}`,
-      },
-      { status: 500 }
-    );
+    // Ensure room_id is a valid ObjectId
+    if (!ObjectId.isValid(room_id)) {
+      return new NextResponse('Invalid Room ID', { status: 400 });
+    }
+
+    console.log("room_id is a valid ObjectId");
+    console.log("Trying to delete room with id of:", room_id);
+
+    const result = await rooms.deleteOne({ _id: new ObjectId(room_id) });
+    console.log("result:", result);
+    if (result.deletedCount === 0) {
+      return new NextResponse(`No room found with id: ${room_id}`, { status: 404 });
+    }
+
+    return new NextResponse(`Room with id of: ${room_id} was successfully deleted`, { status: 200 });
+  } catch (error) {
+    console.error('Failed to connect to the database:', error);
+    return new NextResponse('Database connection failed', { status: 500 });
   }
 }
